apps/nextgen-tm-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nextgen-tm-server"
version = "0.1.0"
description = "FastAPI backend for nextgen-tm-frontend"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2",
  "orjson>=3",
  "httpx>=0.27",
  "fastapi-users[sqlalchemy]>=13.0",
  "aiosqlite>=0.19",
  "google-auth>=2.23",
]

[tool.setuptools.packages.find]
where = ["src"]

